+++ ConnectedGardenPots/ModConfig.cs
@@ -1,7 +1,11 @@
-﻿namespace ConnectedGardenPots
+﻿using StardewModdingAPI;
+using StardewModdingAPI.Utilities;
+
+namespace ConnectedGardenPots
 {
 	public class ModConfig
 	{
 		public bool EnableMod { get; set; } = true;
+		public KeybindList DisconnectKeys { get; set; } = new KeybindList(new Keybind(SButton.LeftControl, SButton.J));
 	}
 }