+++ ConnectedGardenPots/release-notes.md
@@ -2,6 +2,11 @@
 
 # Release notes
 
+## 0.1.1-unofficial.2-mouahrara
+Released on 8 March, 2025, for SMAPI 4.1.0 or later.
+* Added **Disconnect Keys** option ✨
+* Fixed an issue where garden pots were using connected textures even when the textures of the surrounding garden pots had been replaced by the [Alternative Textures](https://www.nexusmods.com/stardewvalley/mods/9246) mod 🔧
+
 ## 0.1.1-unofficial.1-mouahrara
 Released on 4 April, 2024, for SMAPI 4.0.0 or later.
 * Migrated to 1.6.0 and project cleanup 🚀